\name{IDW-package}
\alias{IDW-package}
\alias{IDW}
\docType{package}
\title{
  Inverse Distance Weighting
}
\description{
  Simple and fast functions to conduct Inverse distance weighting 
  interpolation and assess accuracy. Implimented in C++.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{

Maintainer: James L. Martin <contactjamesmartin@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
